+++ rollup.config.js
@@ -1,0 +1,7 @@
+import {rollupOptions} from './tools/rollup_options.js';
+import {uglify} from 'rollup-plugin-uglify';
+
+export default [
+  rollupOptions({}),
+  rollupOptions({suffix: '.min', extraPlugins: [uglify()]}),
+];