+++ src/parsley/validator.js
@@ -56,6 +56,8 @@ Validator.prototype = {
         return this.validateDate(...arguments);
       }
       if (this.validateNumber) {
+        if (!value) // Builtin validators all accept empty strings, except `required` of course
+          return true;
         if (isNaN(value))
           return false;
         arguments[0] = parseFloat(arguments[0]);
@@ -82,7 +84,7 @@ Validator.prototype = {
       for (var i = 0; i < values.length; i++)
         values[i] = Utils.parseRequirement(type[i], values[i]);
       return values;
-    } else if ($.isPlainObject(type)) {
+    } else if (Utils.isPlainObject(type)) {
       return convertExtraOptionRequirement(type, requirements, extraOptionReader);
     } else {
       return [Utils.parseRequirement(type, requirements)];