+++ tools/mocha_options.js
@@ -1,0 +1,9 @@
+export default {
+  globals: [
+    "$",
+    "stub",
+    "spy",
+    "expect",
+    "expectWarning",
+  ],
+};