+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Parsley 2.x changelog
 
+Parsley follows [semver](https://semver.org/). This lists new features only. For bug fixes and small changes, check the commit list.
+
+## 2.9
+
+- Added `euvatin` (European Union Value Added Tax Identification Number) validator.
+
 ## 2.8
 
 - Added `refresh` method. Parsley automatically refreshes before validating, but `refresh` is there if one needs to insure that some fields are rebound/unbound, or triggers are updated, etc...
@@ -7,8 +13,6 @@
 - `Parsley.hasValidator` makes it easy to check if a validator exists or not [#1216]
 
 - `classHandler` now supports global functions, `errorWrapper` too. [#1101]
-
-- Bug fixes...
 
 ## 2.7